security/permissions.py
# permissions.py

from rest_framework.permissions import BasePermission

class IsAdmin(BasePermission):
    def has_permission(self, request, view):
        # Check if the user is an admin
        return request.user.role == 'admin'
